+4XP: Fidgeting; moving types into the world module and renaming some things.
